LPS-90999 Moves to product.navigation.control.menu
