+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/product/navigation/control/menu/CustomizationSettingsProductNavigationControlMenuEntry.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.admin.web.internal.control.menu;
+package com.liferay.layout.admin.web.internal.product.navigation.control.menu;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
