+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/theme/contributor/LayoutCustomizationTemplateContextContributor.java
@@ -14,7 +14,7 @@
 
 package com.liferay.layout.admin.web.internal.theme.contributor;
 
-import com.liferay.layout.admin.web.internal.control.menu.CustomizationSettingsProductNavigationControlMenuEntry;
+import com.liferay.layout.admin.web.internal.product.navigation.control.menu.CustomizationSettingsProductNavigationControlMenuEntry;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
