+++ modules/apps/layout/layout-type-controller/layout-type-controller-content/src/main/java/com/liferay/layout/type/controller/content/internal/product/navigation/control/menu/ToggleEditLayoutModeProductNavigationControlMenuEntry.java
@@ -12,21 +12,23 @@
  * details.
  */
 
-package com.liferay.layout.type.controller.content.internal.control.menu;
+package com.liferay.layout.type.controller.content.internal.product.navigation.control.menu;
 
-import com.liferay.layout.constants.LayoutConstants;
+import com.liferay.layout.content.page.editor.constants.ContentPageEditorWebKeys;
+import com.liferay.layout.page.template.model.LayoutPageTemplateEntry;
+import com.liferay.layout.type.controller.content.internal.controller.ContentLayoutTypeController;
 import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.Layout;
+import com.liferay.portal.kernel.model.LayoutTypeController;
+import com.liferay.portal.kernel.model.LayoutTypePortlet;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.service.permission.LayoutPermissionUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.product.navigation.control.menu.BaseJSPProductNavigationControlMenuEntry;
 import com.liferay.product.navigation.control.menu.ProductNavigationControlMenuEntry;
 import com.liferay.product.navigation.control.menu.constants.ProductNavigationControlMenuCategoryKeys;
 
+import java.util.Locale;
 import java.util.Objects;
 
 import javax.servlet.ServletContext;
@@ -41,18 +43,28 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true,
 	property = {
-		"product.navigation.control.menu.category.key=" + ProductNavigationControlMenuCategoryKeys.TOOLS,
-		"product.navigation.control.menu.entry.order:Integer=100"
+		"product.navigation.control.menu.category.key=" + ProductNavigationControlMenuCategoryKeys.USER,
+		"product.navigation.control.menu.entry.order:Integer=50"
 	},
 	service = ProductNavigationControlMenuEntry.class
 )
-public class LayoutEditorToolbarProductNavigationControlMenuEntry
+public class ToggleEditLayoutModeProductNavigationControlMenuEntry
 	extends BaseJSPProductNavigationControlMenuEntry
 	implements ProductNavigationControlMenuEntry {
 
 	@Override
 	public String getIconJspPath() {
-		return "/layout_editor_toolbar/entry.jsp";
+		return "/entries/toggle_edit_layout_mode.jsp";
+	}
+
+	@Override
+	public String getLabel(Locale locale) {
+		return null;
+	}
+
+	@Override
+	public String getURL(HttpServletRequest request) {
+		return null;
 	}
 
 	@Override
@@ -60,19 +72,26 @@ public class LayoutEditorToolbarProductNavigationControlMenuEntry
 		ThemeDisplay themeDisplay = (ThemeDisplay)request.getAttribute(
 			WebKeys.THEME_DISPLAY);
 
-		Layout layout = themeDisplay.getLayout();
+		LayoutTypePortlet layoutTypePortlet =
+			themeDisplay.getLayoutTypePortlet();
 
-		if (!Objects.equals(
-				layout.getType(), LayoutConstants.LAYOUT_TYPE_ASSET_DISPLAY) &&
-			!Objects.equals(
-				layout.getType(), LayoutConstants.LAYOUT_TYPE_CONTENT)) {
+		LayoutTypeController layoutTypeController =
+			layoutTypePortlet.getLayoutTypeController();
 
+		if (layoutTypeController.isFullPageDisplayable()) {
 			return false;
 		}
 
-		String mode = ParamUtil.getString(request, "p_l_mode", Constants.VIEW);
+		if (!(layoutTypeController instanceof ContentLayoutTypeController)) {
+			return false;
+		}
 
-		if (!Objects.equals(mode, Constants.EDIT)) {
+		String className = (String)request.getAttribute(
+			ContentPageEditorWebKeys.CLASS_NAME);
+
+		if (Objects.equals(
+				className, LayoutPageTemplateEntry.class.getName())) {
+
 			return false;
 		}
